% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HazelcastClient-exports.R
\name{HazelcastClient}
\alias{HazelcastClient}
\title{Create a HazelcastClient Object from the HazelcastClient C++ Class}
\arguments{
\item{name}{Name of HazelcastClient}

\item{age}{Age of HazelcastClient}

\item{male}{Is HazelcastClient a Male?}
}
\value{
A \code{HazelcastClient} object from the \emph{C++} HazelcastClient Class.
}
\description{
Allows for the creation of a HazelcastClient Object in \emph{C++} from \emph{R}
using the \emph{C++} HazelcastClient class.
}
\examples{
##################
## Constructor

# Construct new HazelcastClient object called "ben"
hz = new(HazelcastClient, url = "some url")

##################
## Getters

}
